add TutorCertificateServiceTests

+++ EduMatch.BusinessLogicLayer/Services/TutorCertificateService.cs
@@ -76,6 +76,8 @@ namespace EduMatch.BusinessLogicLayer.Services
 		/// </summary>
 		public async Task<IReadOnlyList<TutorCertificateDto>> GetByCertificateTypeAsync(int certificateTypeId)
 		{
+			if (certificateTypeId <= 0)
+				throw new ArgumentException("CertificateTypeId must be greater than 0");
 			var entities = await _tutorCertificateRepository.GetByCertificateTypeAsync(certificateTypeId);
 			return _mapper.Map<IReadOnlyList<TutorCertificateDto>>(entities);
 		}
@@ -251,6 +253,8 @@ namespace EduMatch.BusinessLogicLayer.Services
 		/// </summary>
 		public async Task DeleteAsync(int id)
 		{
+			if (id <= 0)
+				throw new ArgumentException("ID must be greater than 0");
 			await _tutorCertificateRepository.RemoveByIdAsync(id);
 		}
 
@@ -259,6 +263,8 @@ namespace EduMatch.BusinessLogicLayer.Services
 		/// </summary>
 		public async Task DeleteByTutorIdAsync(int tutorId)
 		{
+			if (tutorId <= 0)
+				throw new ArgumentException("TutorId must be greater than 0");
 			await _tutorCertificateRepository.RemoveByTutorIdAsync(tutorId);
 		}
